Created a stand-alone widget for SEG-Y viewing
